# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Smart Water Pump Monitoring System (SWPMS) API",
    "description": "\n### PT Ascon Multi Pratama - Industrial IoT & Water Pump Monitoring Dashboard\n\nSWPMS is a centralized, real-time, and historical water pump monitoring backend platform.\nThis API provides full operational telemetry queries, dynamic site/area/device/sensor topology management, alarm rule evaluation, audit logging, CSV reporting, and controlled commanded pump ON/OFF power operations.\n\n#### Authentication\nUse the `/api/v1/auth/login` endpoint to obtain a JWT Bearer token.\nProvide the token in the header:\n`Authorization: Bearer <your_jwt_token>`\n\n#### Postman Import Instructions\nYou can import this entire API into Postman directly by choosing **Import -> Link** and pasting:\n`http://localhost:3000/docs/json` or downloading the OpenAPI specification from `/docs/json`.\n        ",
    "version": "1.1.0",
    "contact": {
      "name": "PT Ascon Multi Pratama SCADA & Automation Team",
      "url": "https://ascon.co.id"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /api/v1/auth/login"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/reports/telemetry.csv": {
      "get": {
        "summary": "Export Telemetry Samples to CSV",
        "tags": [
          "Reports"
        ],
        "description": "Streams historical telemetry data as CSV file with custom date range and asset filters.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "site_id",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "asset_id",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "from",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "to",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/reports/alarms.csv": {
      "get": {
        "summary": "Export Alarms to CSV",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "site_id",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/reports/events.csv": {
      "get": {
        "summary": "Export Events to CSV",
        "tags": [
          "Reports"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "site_id",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
...
}